+++ packages/router/router.ts
@@ -3,7 +3,8 @@ import {
   Action as HistoryAction,
   createLocation,
   createPath,
-  createURL,
+  createClientSideURL,
+  invariant,
   parsePath,
 } from "./history";
 import type {
@@ -20,7 +21,7 @@ import type {
   Submission,
   SuccessResult,
   AgnosticRouteMatch,
-  SubmissionFormMethod,
+  MutationFormMethod,
 } from "./utils";
 import {
   DeferredData,
@@ -28,7 +29,6 @@ import {
   ResultType,
   convertRoutesToDataRoutes,
   getPathContributingMatches,
-  invariant,
   isRouteErrorResponse,
   joinPaths,
   matchRoutes,
@@ -380,8 +380,14 @@ export interface StaticHandlerContext {
  */
 export interface StaticHandler {
   dataRoutes: AgnosticDataRouteObject[];
-  query(request: Request): Promise<StaticHandlerContext | Response>;
-  queryRoute(request: Request, routeId?: string): Promise<any>;
+  query(
+    request: Request,
+    opts?: { requestContext?: unknown }
+  ): Promise<StaticHandlerContext | Response>;
+  queryRoute(
+    request: Request,
+    opts?: { routeId?: string; requestContext?: unknown }
+  ): Promise<any>;
 }
 
 /**
@@ -591,15 +597,20 @@ interface QueryRouteResponse {
   response: Response;
 }
 
-const validActionMethodsArr: SubmissionFormMethod[] = [
+const validMutationMethodsArr: MutationFormMethod[] = [
   "post",
   "put",
   "patch",
   "delete",
 ];
-const validActionMethods = new Set<SubmissionFormMethod>(validActionMethodsArr);
+const validMutationMethods = new Set<MutationFormMethod>(
+  validMutationMethodsArr
+);
 
-const validRequestMethodsArr: FormMethod[] = ["get", ...validActionMethodsArr];
+const validRequestMethodsArr: FormMethod[] = [
+  "get",
+  ...validMutationMethodsArr,
+];
 const validRequestMethods = new Set<FormMethod>(validRequestMethodsArr);
 
 const redirectStatusCodes = new Set([301, 302, 303, 307, 308]);
@@ -979,7 +990,8 @@ export function createRouter(init: RouterInit): Router {
     };
 
     let historyAction =
-      (opts && opts.replace) === true || submission != null
+      (opts && opts.replace) === true ||
+      (submission != null && isMutationMethod(submission.formMethod))
         ? HistoryAction.Replace
         : HistoryAction.Push;
     let preventScrollReset =
@@ -1087,7 +1099,7 @@ export function createRouter(init: RouterInit): Router {
 
     // Create a controller/Request for this navigation
     pendingNavigationController = new AbortController();
-    let request = createRequest(
+    let request = createClientSideRequest(
       location,
       pendingNavigationController.signal,
       opts && opts.submission
@@ -1103,7 +1115,11 @@ export function createRouter(init: RouterInit): Router {
       pendingError = {
         [findNearestBoundary(matches).route.id]: opts.pendingError,
       };
-    } else if (opts && opts.submission) {
+    } else if (
+      opts &&
+      opts.submission &&
+      isMutationMethod(opts.submission.formMethod)
+    ) {
       // Call action if we received an action submission
       let actionOutput = await handleAction(
         request,
@@ -1128,7 +1144,7 @@ export function createRouter(init: RouterInit): Router {
       loadingNavigation = navigation;
 
       // Create a GET request for the loaders
-      request = createRequest(request.url, request.signal);
+      request = new Request(request.url, { signal: request.signal });
     }
 
     // Call loaders
@@ -1263,6 +1279,7 @@ export function createRouter(init: RouterInit): Router {
         formAction: undefined,
         formEncType: undefined,
         formData: undefined,
+        ...submission,
       };
       loadingNavigation = navigation;
     }
@@ -1427,7 +1444,7 @@ export function createRouter(init: RouterInit): Router {
     let { path, submission } = normalizeNavigateOptions(href, opts, true);
     let match = getTargetMatch(matches, path);
 
-    if (submission) {
+    if (submission && isMutationMethod(submission.formMethod)) {
       handleFetcherAction(key, routeId, path, match, matches, submission);
       return;
     }
@@ -1435,7 +1452,7 @@ export function createRouter(init: RouterInit): Router {
     // Store off the match so we can call it's shouldRevalidate on subsequent
     // revalidations
     fetchLoadMatches.set(key, [path, match, matches]);
-    handleFetcherLoader(key, routeId, path, match, matches);
+    handleFetcherLoader(key, routeId, path, match, matches, submission);
   }
 
   // Call the action for the matched fetcher.submit(), and then handle redirects,
@@ -1473,7 +1490,11 @@ export function createRouter(init: RouterInit): Router {
 
     // Call the action for the fetcher
     let abortController = new AbortController();
-    let fetchRequest = createRequest(path, abortController.signal, submission);
+    let fetchRequest = createClientSideRequest(
+      path,
+      abortController.signal,
+      submission
+    );
     fetchControllers.set(key, abortController);
 
     let actionResult = await callLoaderOrAction(
@@ -1520,7 +1541,7 @@ export function createRouter(init: RouterInit): Router {
     // Start the data load for current matches, or the next location if we're
     // in the middle of a navigation
     let nextLocation = state.navigation.location || state.location;
-    let revalidationRequest = createRequest(
+    let revalidationRequest = createClientSideRequest(
       nextLocation,
       abortController.signal
     );
@@ -1658,7 +1679,8 @@ export function createRouter(init: RouterInit): Router {
     routeId: string,
     path: string,
     match: AgnosticDataRouteMatch,
-    matches: AgnosticDataRouteMatch[]
+    matches: AgnosticDataRouteMatch[],
+    submission?: Submission
   ) {
     let existingFetcher = state.fetchers.get(key);
     // Put this fetcher into it's loading state
@@ -1668,6 +1690,7 @@ export function createRouter(init: RouterInit): Router {
       formAction: undefined,
       formEncType: undefined,
       formData: undefined,
+      ...submission,
       data: existingFetcher && existingFetcher.data,
     };
     state.fetchers.set(key, loadingFetcher);
@@ -1675,7 +1698,7 @@ export function createRouter(init: RouterInit): Router {
 
     // Call the loader for this fetcher route match
     let abortController = new AbortController();
-    let fetchRequest = createRequest(path, abortController.signal);
+    let fetchRequest = createClientSideRequest(path, abortController.signal);
     fetchControllers.set(key, abortController);
     let result: DataResult = await callLoaderOrAction(
       "loader",
@@ -1776,17 +1799,17 @@ export function createRouter(init: RouterInit): Router {
       "Expected a location on the redirect navigation"
     );
 
-    if (
-      redirect.external &&
-      typeof window !== "undefined" &&
-      typeof window.location !== "undefined"
-    ) {
-      if (replace) {
-        window.location.replace(redirect.location);
-      } else {
-        window.location.assign(redirect.location);
+    // Check if this an external redirect that goes to a new origin
+    if (typeof window?.location !== "undefined") {
+      let newOrigin = createClientSideURL(redirect.location).origin;
+      if (window.location.origin !== newOrigin) {
+        if (replace) {
+          window.location.replace(redirect.location);
+        } else {
+          window.location.assign(redirect.location);
+        }
+        return;
       }
-      return;
     }
 
     // There's no need to abort on redirects, since we don't detect the
@@ -1799,12 +1822,12 @@ export function createRouter(init: RouterInit): Router {
     let { formMethod, formAction, formEncType, formData } = state.navigation;
 
     // If this was a 307/308 submission we want to preserve the HTTP method and
-    // re-submit the POST/PUT/PATCH/DELETE as a submission navigation to the
+    // re-submit the GET/POST/PUT/PATCH/DELETE as a submission navigation to the
     // redirected location
     if (
       redirectPreserveMethodStatusCodes.has(redirect.status) &&
       formMethod &&
-      isSubmissionMethod(formMethod) &&
+      isMutationMethod(formMethod) &&
       formEncType &&
       formData
     ) {
@@ -1849,7 +1872,7 @@ export function createRouter(init: RouterInit): Router {
       ...fetchersToLoad.map(([, href, match, fetchMatches]) =>
         callLoaderOrAction(
           "loader",
-          createRequest(href, request.signal),
+          createClientSideRequest(href, request.signal),
           match,
           fetchMatches,
           router.basename
@@ -2118,7 +2141,8 @@ export function unstable_createStaticHandler(
    * return it directly.
    */
   async function query(
-    request: Request
+    request: Request,
+    { requestContext }: { requestContext?: unknown } = {}
   ): Promise<StaticHandlerContext | Response> {
     let url = new URL(request.url);
     let method = request.method.toLowerCase();
@@ -2162,8 +2186,8 @@ export function unstable_createStaticHandler(
       };
     }
 
-    let result = await queryImpl(request, location, matches);
-    if (result instanceof Response) {
+    let result = await queryImpl(request, location, matches, requestContext);
+    if (isResponse(result)) {
       return result;
     }
 
@@ -2193,7 +2217,13 @@ export function unstable_createStaticHandler(
    * code.  Examples here are 404 and 405 errors that occur prior to reaching
    * any user-defined loaders.
    */
-  async function queryRoute(request: Request, routeId?: string): Promise<any> {
+  async function queryRoute(
+    request: Request,
+    {
+      routeId,
+      requestContext,
+    }: { requestContext?: unknown; routeId?: string } = {}
+  ): Promise<any> {
     let url = new URL(request.url);
     let method = request.method.toLowerCase();
     let location = createLocation("", createPath(url), null, "default");
@@ -2220,8 +2250,14 @@ export function unstable_createStaticHandler(
       throw getInternalRouterError(404, { pathname: location.pathname });
     }
 
-    let result = await queryImpl(request, location, matches, match);
-    if (result instanceof Response) {
+    let result = await queryImpl(
+      request,
+      location,
+      matches,
+      requestContext,
+      match
+    );
+    if (isResponse(result)) {
       return result;
     }
 
@@ -2243,6 +2279,7 @@ export function unstable_createStaticHandler(
     request: Request,
     location: Location,
     matches: AgnosticDataRouteMatch[],
+    requestContext: unknown,
     routeMatch?: AgnosticDataRouteMatch
   ): Promise<Omit<StaticHandlerContext, "location" | "basename"> | Response> {
     invariant(
@@ -2251,18 +2288,24 @@ export function unstable_createStaticHandler(
     );
 
     try {
-      if (isSubmissionMethod(request.method.toLowerCase())) {
+      if (isMutationMethod(request.method.toLowerCase())) {
         let result = await submit(
           request,
           matches,
           routeMatch || getTargetMatch(matches, location),
+          requestContext,
           routeMatch != null
         );
         return result;
       }
 
-      let result = await loadRouteData(request, matches, routeMatch);
-      return result instanceof Response
+      let result = await loadRouteData(
+        request,
+        matches,
+        requestContext,
+        routeMatch
+      );
+      return isResponse(result)
         ? result
         : {
             ...result,
@@ -2292,6 +2335,7 @@ export function unstable_createStaticHandler(
     request: Request,
     matches: AgnosticDataRouteMatch[],
     actionMatch: AgnosticDataRouteMatch,
+    requestContext: unknown,
     isRouteRequest: boolean
   ): Promise<Omit<StaticHandlerContext, "location" | "basename"> | Response> {
     let result: DataResult;
@@ -2299,7 +2343,7 @@ export function unstable_createStaticHandler(
     if (!actionMatch.route.action) {
       let error = getInternalRouterError(405, {
         method: request.method,
-        pathname: createURL(request.url).pathname,
+        pathname: new URL(request.url).pathname,
         routeId: actionMatch.route.id,
       });
       if (isRouteRequest) {
@@ -2317,7 +2361,8 @@ export function unstable_createStaticHandler(
         matches,
         basename,
         true,
-        isRouteRequest
+        isRouteRequest,
+        requestContext
       );
 
       if (request.signal.aborted) {
@@ -2367,9 +2412,15 @@ export function unstable_createStaticHandler(
       // Store off the pending error - we use it to determine which loaders
       // to call and will commit it when we complete the navigation
       let boundaryMatch = findNearestBoundary(matches, actionMatch.route.id);
-      let context = await loadRouteData(request, matches, undefined, {
-        [boundaryMatch.route.id]: result.error,
-      });
+      let context = await loadRouteData(
+        request,
+        matches,
+        requestContext,
+        undefined,
+        {
+          [boundaryMatch.route.id]: result.error,
+        }
+      );
 
       // action status codes take precedence over loader status codes
       return {
@@ -2385,8 +2436,8 @@ export function unstable_createStaticHandler(
     }
 
     // Create a GET request for the loaders
-    let loaderRequest = createRequest(request.url, request.signal);
-    let context = await loadRouteData(loaderRequest, matches);
+    let loaderRequest = new Request(request.url, { signal: request.signal });
+    let context = await loadRouteData(loaderRequest, matches, requestContext);
 
     return {
       ...context,
@@ -2404,6 +2455,7 @@ export function unstable_createStaticHandler(
   async function loadRouteData(
     request: Request,
     matches: AgnosticDataRouteMatch[],
+    requestContext: unknown,
     routeMatch?: AgnosticDataRouteMatch,
     pendingActionError?: RouteData
   ): Promise<
@@ -2419,7 +2471,7 @@ export function unstable_createStaticHandler(
     if (isRouteRequest && !routeMatch?.route.loader) {
       throw getInternalRouterError(400, {
         method: request.method,
-        pathname: createURL(request.url).pathname,
+        pathname: new URL(request.url).pathname,
         routeId: routeMatch?.route.id,
       });
     }
@@ -2452,7 +2504,8 @@ export function unstable_createStaticHandler(
           matches,
           basename,
           true,
-          isRouteRequest
+          isRouteRequest,
+          requestContext
         )
       ),
     ]);
@@ -2548,17 +2601,19 @@ function normalizeNavigateOptions(
   }
 
   // Create a Submission on non-GET navigations
-  if (opts.formMethod && isSubmissionMethod(opts.formMethod)) {
-    return {
-      path,
-      submission: {
-        formMethod: opts.formMethod,
-        formAction: stripHashFromPath(path),
-        formEncType:
-          (opts && opts.formEncType) || "application/x-www-form-urlencoded",
-        formData: opts.formData,
-      },
+  let submission: Submission | undefined;
+  if (opts.formData) {
+    submission = {
+      formMethod: opts.formMethod || "get",
+      formAction: stripHashFromPath(path),
+      formEncType:
+        (opts && opts.formEncType) || "application/x-www-form-urlencoded",
+      formData: opts.formData,
     };
+
+    if (isMutationMethod(submission.formMethod)) {
+      return { path, submission };
+    }
   }
 
   // Flatten submission onto URLSearchParams for GET submissions
@@ -2583,7 +2638,7 @@ function normalizeNavigateOptions(
     };
   }
 
-  return { path: createPath(parsedPath) };
+  return { path: createPath(parsedPath), submission };
 }
 
 // Filter out all routes below any caught error as they aren't going to
@@ -2710,9 +2765,9 @@ function shouldRevalidateLoader(
   isRevalidationRequired: boolean,
   actionResult: DataResult | undefined
 ) {
-  let currentUrl = createURL(currentLocation);
+  let currentUrl = createClientSideURL(currentLocation);
   let currentParams = currentMatch.params;
-  let nextUrl = createURL(location);
+  let nextUrl = createClientSideURL(location);
   let nextParams = match.params;
 
   // This is the default implementation as to when we revalidate.  If the route
@@ -2755,7 +2810,8 @@ async function callLoaderOrAction(
   matches: AgnosticDataRouteMatch[],
   basename = "/",
   isStaticRequest: boolean = false,
-  isRouteRequest: boolean = false
+  isRouteRequest: boolean = false,
+  requestContext?: unknown
 ): Promise<DataResult> {
   let resultType;
   let result;
@@ -2774,7 +2830,7 @@ async function callLoaderOrAction(
     );
 
     result = await Promise.race([
-      handler({ request, params: match.params }),
+      handler({ request, params: match.params, context: requestContext }),
       abortPromise,
     ]);
 
@@ -2791,7 +2847,7 @@ async function callLoaderOrAction(
     request.signal.removeEventListener("abort", onReject);
   }
 
-  if (result instanceof Response) {
+  if (isResponse(result)) {
     let status = result.status;
 
     // Process redirects
@@ -2802,17 +2858,20 @@ async function callLoaderOrAction(
         "Redirects returned/thrown from loaders/actions must have a Location header"
       );
 
-      // Check if this an external redirect that goes to a new origin
-      let external = createURL(location).origin !== createURL("/").origin;
+      let isAbsolute =
+        /^[a-z+]+:\/\//i.test(location) || location.startsWith("//");
 
       // Support relative routing in internal redirects
-      if (!external) {
+      if (!isAbsolute) {
         let activeMatches = matches.slice(0, matches.indexOf(match) + 1);
         let routePathnames = getPathContributingMatches(activeMatches).map(
           (match) => match.pathnameBase
         );
-        let requestPath = createURL(request.url).pathname;
-        let resolvedLocation = resolveTo(location, routePathnames, requestPath);
+        let resolvedLocation = resolveTo(
+          location,
+          routePathnames,
+          new URL(request.url).pathname
+        );
         invariant(
           createPath(resolvedLocation),
           `Unable to resolve redirect location: ${location}`
@@ -2842,7 +2901,6 @@ async function callLoaderOrAction(
         status,
         location,
         revalidate: result.headers.get("X-Remix-Revalidate") !== null,
-        external,
       };
     }
 
@@ -2892,15 +2950,18 @@ async function callLoaderOrAction(
   return { type: ResultType.data, data: result };
 }
 
-function createRequest(
+// Utility method for creating the Request instances for loaders/actions during
+// client-side navigations and fetches.  During SSR we will always have a
+// Request instance from the static handler (query/queryRoute)
+function createClientSideRequest(
   location: string | Location,
   signal: AbortSignal,
   submission?: Submission
 ): Request {
-  let url = createURL(stripHashFromPath(location)).toString();
+  let url = createClientSideURL(stripHashFromPath(location)).toString();
   let init: RequestInit = { signal };
 
-  if (submission) {
+  if (submission && isMutationMethod(submission.formMethod)) {
     let { formMethod, formEncType, formData } = submission;
     init.method = formMethod.toUpperCase();
     init.body =
@@ -3138,12 +3199,10 @@ function getInternalRouterError(
     pathname,
     routeId,
     method,
-    message,
   }: {
     pathname?: string;
     routeId?: string;
     method?: string;
-    message?: string;
   } = {}
 ) {
   let statusText = "Unknown Server Error";
@@ -3218,8 +3277,18 @@ function isRedirectResult(result?: DataResult): result is RedirectResult {
   return (result && result.type) === ResultType.redirect;
 }
 
+function isResponse(value: any): value is Response {
+  return (
+    value != null &&
+    typeof value.status === "number" &&
+    typeof value.statusText === "string" &&
+    typeof value.headers === "object" &&
+    typeof value.body !== "undefined"
+  );
+}
+
 function isRedirectResponse(result: any): result is Response {
-  if (!(result instanceof Response)) {
+  if (!isResponse(result)) {
     return false;
   }
 
@@ -3231,7 +3300,7 @@ function isRedirectResponse(result: any): result is Response {
 function isQueryRouteResponse(obj: any): obj is QueryRouteResponse {
   return (
     obj &&
-    obj.response instanceof Response &&
+    isResponse(obj.response) &&
     (obj.type === ResultType.data || ResultType.error)
   );
 }
@@ -3240,8 +3309,8 @@ function isValidMethod(method: string): method is FormMethod {
   return validRequestMethods.has(method as FormMethod);
 }
 
-function isSubmissionMethod(method: string): method is SubmissionFormMethod {
-  return validActionMethods.has(method as SubmissionFormMethod);
+function isMutationMethod(method?: string): method is MutationFormMethod {
+  return validMutationMethods.has(method as MutationFormMethod);
 }
 
 async function resolveDeferredResults(